Break player data out from main CandyStore class
